Revert "refactored"
This reverts commit 813a987cd7e7db4269b42dce83ba74a388186af5.

+++ src/main/java/com/gmail/spanteleyko/web/repositories/RoleRepository.java
@@ -7,7 +7,7 @@ import java.sql.SQLException;
 import java.util.List;
 
 public interface RoleRepository {
-    List<Role> get(Connection connection, Long userId) throws SQLException;
+    List<Role> get(Connection connection, int userId) throws SQLException;
 
     List<Role> get(Connection connection) throws SQLException;
